+++ week3/assignment/점메추 🍚/src/components/MainCharacter.jsx
