+++ week3/assignment/점메추 🍚/src/components/ProgressBar.jsx
